+++ test/integrationTests/advisor.integration.test.ts
@@ -5,9 +5,9 @@
 
 import * as vscode from 'vscode';
 
-import { expect } from 'chai';
+import { expect, should } from 'chai';
 import * as path from 'path';
-import { activateCSharpExtension, isRazorWorkspace } from './integrationHelpers';
+import { activateCSharpExtension, isRazorWorkspace, isSlnWithGenerator } from './integrationHelpers';
 import testAssetWorkspace from './testAssets/testAssetWorkspace';
 
 import { Advisor } from '../../src/features/diagnosticsProvider';
@@ -25,8 +25,9 @@ suite(`Advisor ${testAssetWorkspace.description}`, function () {
     let advisor: Advisor;
 
     suiteSetup(async function () {
-        // These tests don't run on the BasicRazorApp2_1 solution
-        if (isRazorWorkspace(vscode.workspace)) {
+        should();
+
+        if (isRazorWorkspace(vscode.workspace) || isSlnWithGenerator(vscode.workspace)) {
             this.skip();
         }