+++ test/integrationTests/documentSymbolProvider.integration.test.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode';
 import * as path from 'path';
 
 import { should, expect } from 'chai';
-import { activateCSharpExtension, isRazorWorkspace } from './integrationHelpers';
+import * as integrationHelpers from './integrationHelpers';
 import testAssetWorkspace from './testAssets/testAssetWorkspace';
 
 const chai = require('chai');
@@ -20,12 +20,11 @@ suite(`DocumentSymbolProvider: ${testAssetWorkspace.description}`, function () {
     suiteSetup(async function () {
         should();
 
-        // These tests don't run on the BasicRazorApp2_1 solution
-        if (isRazorWorkspace(vscode.workspace)) {
+        if (integrationHelpers.isRazorWorkspace(vscode.workspace) || integrationHelpers.isSlnWithGenerator(vscode.workspace)) {
             this.skip();
         }
 
-        await activateCSharpExtension();
+        await integrationHelpers.activateCSharpExtension();
         await testAssetWorkspace.restore();
 
         let fileName = 'documentSymbols.cs';