+++ test/integrationTests/semanticTokensProvider.test.ts
@@ -7,10 +7,8 @@ import * as vscode from 'vscode';
 import * as path from 'path';
 
 import { should, assert } from 'chai';
-import { activateCSharpExtension, isRazorWorkspace } from './integrationHelpers';
+import { activateCSharpExtension, isRazorWorkspace, isSlnWithGenerator } from './integrationHelpers';
 import testAssetWorkspace from './testAssets/testAssetWorkspace';
-import { EventType } from '../../src/omnisharp/EventType';
-import { poll } from './poll';
 
 const chai = require('chai');
 chai.use(require('chai-arrays'));
@@ -25,10 +23,10 @@ interface ExpectedToken {
 
 async function assertTokens(fileUri: vscode.Uri, expected: ExpectedToken[] | null, message?: string): Promise<void> {
 
-    const legend = <vscode.SemanticTokensLegend>await vscode.commands.executeCommand("csharp.private.getSemanticTokensLegend");
-    const actual = <vscode.SemanticTokens>await vscode.commands.executeCommand("csharp.private.getSemanticTokens", fileUri);
+    const legend = <vscode.SemanticTokensLegend>await vscode.commands.executeCommand("vscode.provideDocumentSemanticTokensLegend", fileUri);
+    const actual = <vscode.SemanticTokens>await vscode.commands.executeCommand("vscode.provideDocumentSemanticTokens", fileUri);
 
-    if (actual === null) {
+    if (!actual) {
         assert.isNull(expected, message);
         return;
     }
@@ -54,30 +52,18 @@ suite(`SemanticTokensProvider: ${testAssetWorkspace.description}`, function () {
     suiteSetup(async function () {
         should();
 
-        // These tests don't run on the BasicRazorApp2_1 solution
-        if (isRazorWorkspace(vscode.workspace)) {
+        if (isRazorWorkspace(vscode.workspace) || isSlnWithGenerator(vscode.workspace)) {
             this.skip();
         }
 
         const activation = await activateCSharpExtension();
-        await testAssetWorkspace.restore();
-
-        // Wait for workspace information to be returned
-        let isWorkspaceLoaded = false;
-
-        const subscription = activation.eventStream.subscribe(event => {
-            if (event.type === EventType.WorkspaceInformationUpdated) {
-                isWorkspaceLoaded = true;
-                subscription.unsubscribe();
-            }
-        });
-
-        await poll(() => isWorkspaceLoaded, 25000, 500);
+        await testAssetWorkspace.restoreAndWait(activation);
 
         const fileName = 'semantictokens.cs';
         const projectDirectory = testAssetWorkspace.projects[0].projectDirectoryPath;
 
         fileUri = vscode.Uri.file(path.join(projectDirectory, fileName));
+        await vscode.commands.executeCommand("vscode.open", fileUri);
     });
 
     test('Semantic Highlighting returns null when disabled', async () => {