+++ test/unitTests/optionStream.test.ts
@@ -55,6 +55,7 @@ suite('OptionStream', () => {
             options.enableEditorConfigSupport.should.equal(false);
             options.enableDecompilationSupport.should.equal(false);
             options.enableImportCompletion.should.equal(false);
+            options.enableAsyncCompletion.should.equal(false);
             expect(options.defaultLaunchSolution).to.be.undefined;
         });